[#216] move Repositories into Repository
Also no longer default Elasticsearch index.

Also namespace Elasticsearch test indexes.
