+++ lib/register_sources_bods/ingester/ingest_bulk.rb
@@ -8,7 +8,7 @@ require_relative '../config/adapters'
 require_relative '../logging'
 require_relative '../record_deserializer'
 require_relative '../record_serializer'
-require_relative '../repositories/bods_statement_repository'
+require_relative '../repository'
 require_relative '../services/es_index_creator'
 
 module RegisterSourcesBods
@@ -48,13 +48,14 @@ module RegisterSourcesBods
           serializer: RecordSerializer.new
         ))
         @deserializer = RecordDeserializer.new
-        @repository = repository || Repositories::BodsStatementRepository.new(index:)
-        @es_index_creator = es_index_creator || Services::EsIndexCreator.new(index:)
+        @repository = repository || Repository.new(index:)
+        @es_index_creator = es_index_creator || Services::EsIndexCreator.new
+        @index = index
       end
       # rubocop:enable Metrics/ParameterLists
 
       def call(s3_prefix)
-        es_index_creator.create_index_unless_exists
+        es_index_creator.create_index_unless_exists(@index)
 
         bulk_transformer.call(s3_prefix) do |rows|
           process_rows rows
