+++ lib/register_sources_bods/services/pending_records.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../constants/identifiers'
-require_relative '../repositories/bods_statement_repository'
+require_relative '../repository'
 require_relative 'builder'
 
 module RegisterSourcesBods
@@ -10,7 +10,7 @@ module RegisterSourcesBods
       PreprocessedRecord = Struct.new(:uid, :record, :identifiers, :source)
 
       def initialize(repository: nil, builder: nil)
-        @repository = repository || RegisterSourcesBods::Repositories::BodsStatementRepository.new(
+        @repository = repository || RegisterSourcesBods::Repository.new(
           client: RegisterSourcesBods::Config::ELASTICSEARCH_CLIENT
         )
         @builder = builder || Services::Builder.new
